+++ game-frontend/CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # OWGE changelog
 
-v0.11.1 (latest)
+v0.11.1 (2023-08-14 18:49)
 =================
 * __Improvement:__ [Show faction name of the user in the ranking #528](https://github.com/KevinGuancheDarias/owge/issues/528)  
 * __Fix:__ [System messages not working #522](https://github.com/KevinGuancheDarias/owge/issues/522)